strongbox-authentication/strongbox-authentication-ldap-impls/strongbox-authentication-ldap-core/src/main/java/org/carlspring/strongbox/xml/parsers/LDAPConfigurationParser.java
package org.carlspring.strongbox.xml.parsers;

import org.carlspring.strongbox.configuration.AuthenticationConfiguration;
import org.carlspring.strongbox.configuration.LDAPConfiguration;

import com.thoughtworks.xstream.XStream;

/**
 * @author mtodorov
 */
public class LDAPConfigurationParser
        extends GenericParser<LDAPConfiguration>
{

    public XStream getXStreamInstance()
    {
        XStream xstream = new XStream();
        xstream.autodetectAnnotations(true);
        xstream.alias("ldap-configuration", LDAPConfiguration.class);

        return xstream;
    }

}
